+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Add command `enum:to-native` to convert a class that extends `BenSampo\Enum\Enum` to a native PHP enum
 
+### Fixed
+
+- Prevent translations to be discarded [324](https://github.com/BenSampo/laravel-enum/pull/324)
+
 ## 6.3.3
 
 ### Fixed